+++ src/main/php/predicate/predicates.php
@@ -112,37 +112,24 @@ function isFalse(): Predicate
  * returns predicate which tests for equality
  *
  * @api
- * @deprecated  param  $delta is deprecated, use withDelta() on returned Predicate instead, will be removed with 7.0
- * @param   mixed  $expected  expected value
- * @param   float  $delta     optional  allowed numerical distance between two values to consider them equal
- * @return  \bovigo\assert\predicate\Equals
  */
-function equals($expected, float $delta = 0.0): Equals
+function equals($expected): Equals
 {
-    return new Equals($expected, $delta);
+    return new Equals($expected);
 }
 
 /**
  * returns predicate which tests for non-equality
  *
  * @api
- * @deprecated  param  $delta is deprecated, use withDelta() on returned Predicate instead, will be removed with 7.0
- * @param   mixed  $unexpected  expected value
- * @param   float  $delta       optional  allowed numerical distance between two values to consider them not equal
- * @return  \bovigo\assert\predicate\Delta&\bovigo\assert\predicate\Predicate
  */
-function isNotEqualTo($unexpected, float $delta = 0.0): Delta
+function isNotEqualTo($unexpected): Delta
 {
-    return new class(equals($unexpected, $delta)) extends NegatePredicate implements Delta
+    return new class(equals($unexpected)) extends NegatePredicate implements Delta
     {
-        /**
-         * @var  Equals
-         */
-        private $equal;
-        public function __construct(Equals $equal)
+        public function __construct(private Equals $equal)
         {
-            $this->equal = $equal;
-            parent::__construct($equal);
+            parent::__construct($this->equal);
         }
 
         public function withDelta(float $delta): Predicate